requirements.txt
fastapi
uvicorn
python-socketio
python-multipart
# Google GenAI SDK (v1beta)
google-genai
# Computer Vision & Audio
opencv-python
pyaudio
pillow
mss
# Browser Automation
playwright
# Smart Home
python-kasa
# 3D Printing
zeroconf>=0.131.0
aiohttp>=3.9.0
# Utilities
python-dotenv
# Face & Hand tracking
mediapipe
# CAD Generation
build123d
# Google Workspace Integration
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
